finished up the refractive index finder
given we find

m - number of fringes
lam - the wavelength of the laser
d - the thickness of the slab we are testing
phi_i - the angle to the normal through which we rotated the slab

we will find the refractive index for the material.

I took the liberty to test some values from the report we looked at for inspiration and they are accurate to the digits presented in the report. :)